Traitement des données envoyées par le MovementSerializer

+++ backend/src/classification/Step.java
@@ -11,13 +11,13 @@ public class Step implements Serializable
 	private static final long serialVersionUID = -7057741325158870422L;
 
 	private long time;
-	private float[][] coordinates;
+	private double[][] coordinates;
 	
 	public Step()
 	{
 	}
 	
-	public void setJoint(int joint, float x, float y, float z)
+	public void setJoint(int joint, double x, double y, double z)
 	{
 		coordinates[joint][0] = x;
 		coordinates[joint][1] = y;
@@ -29,7 +29,7 @@ public class Step implements Serializable
 		this.time = time;
 	}
 
-	public float[][] getCoordinates() {
+	public double[][] getCoordinates() {
 		return coordinates;
 	}
